fix: fixed App component
