+++ my-app/package.json
@@ -1,7 +1,7 @@
 {
   "name": "my-app",
   "version": "0.1.0",
-  "homepage": "https://mardon07.github.io/",
+  "homepage": "https://mardon07.github.io/search-book",
   "private": true,
   "dependencies": {
     "@mui/icons-material": "^5.14.8",
