+++ my-app/src/store/reducers/books.slice.ts
